Change line count from int to long
This will avoid potenital issues for files that concatenate to be
very exceptionally large evaluations by increasing the maximum lines
to 2^64

+++ RPN.java
@@ -11,7 +11,7 @@ public class RPN {
 
         HashMap<String,BigInteger> variableMap = new HashMap<String, BigInteger>();
         boolean quitSignal = false;
-        int lineNum = 1;
+        long lineNum = 1;
         String input;
 
         while (!quitSignal) {
